+++ src/index.ts
@@ -30,10 +30,9 @@ export {
   getPrevState,
   getRefs,
   getState,
-  getState_OLD,
   removeItem,
   setState,
-  setState_OLD,
+  setNestedState,
   onNextTick,
   getItemWillBeAdded,
   getItemWillBeRemoved,
