+++ src/usable/getSet.ts
@@ -1,4 +1,4 @@
-import { _addItem, _removeItem, _setState, _setState_OLD } from "../helpers/setting";
+import { forEach } from "chootils/dist/loops";
 import { repondMeta as meta } from "../meta";
 import {
   AllRefs,
@@ -12,6 +12,121 @@ import {
   RepondCallback,
 } from "../types";
 import { applyPatch, getPatch } from "../usable/patchesAndDiffs";
+import { runWhenAddingAndRemovingItems, whenSettingStates } from "../helpers/runWhens";
+import { mergeToState } from "../copyStates";
+
+export function setState(propPath: string, newValue: any, itemId?: string) {
+  whenSettingStates(() => {
+    if (newValue === undefined) return;
+
+    if (!propPath) {
+      console.error("propPath must be provided");
+      return;
+    }
+
+    if (typeof propPath !== "string") {
+      console.error("propPath must be a string");
+      console.log("propPath", propPath);
+
+      return;
+    }
+
+    const storeType = meta.itemTypeByPropPathId[propPath];
+    const propKey = meta.propKeyByPropPathId[propPath];
+    let foundItemId = itemId || meta.itemIdsByItemType[storeType]?.[0];
+    if (!foundItemId) {
+      foundItemId = Object.keys(meta.nowState[storeType] ?? {})[0];
+      console.warn(
+        `${propPath}No itemId found for setState ${storeType}, using first found itemId: ${foundItemId} from Object keys`
+      );
+    }
+
+    mergeToState(
+      storeType,
+      propKey,
+      newValue,
+      foundItemId,
+      meta.nowState,
+      meta.nowMetaPhase === "runningEffects" ? meta.recordedEffectChanges : meta.recordedStepEndEffectChanges,
+      meta.recordedStepEndEffectChanges
+    );
+  });
+}
+
+export function setNestedState(newState: any) {
+  whenSettingStates(() => {
+    if (!newState) return;
+    const itemTypes = Object.keys(newState);
+    forEach(itemTypes, (itemType) => {
+      const itemIds = Object.keys(newState[itemType]);
+      forEach(itemIds, (itemId) => {
+        const itemProps = Object.keys(newState[itemType][itemId]);
+        forEach(itemProps, (propName) => {
+          const newValue = newState[itemType][itemId][propName];
+          setState(`${itemType}.${propName}`, newValue, itemId);
+        });
+      });
+    });
+  });
+}
+
+export function _removeItem({ type, id }: { type: string; id: string }, callback?: any) {
+  if (!meta.willRemoveItemsInfo[type]) meta.willRemoveItemsInfo[type] = {};
+  meta.willRemoveItemsInfo[type][id] = true;
+  runWhenAddingAndRemovingItems(() => {
+    // removing itemId
+    delete meta.nowState[type][id];
+    meta.itemIdsByItemType[type] = Object.keys(meta.nowState[type]);
+    // delete meta.currentRefs[itemType][itemId]; // now done at the end of update repond
+    meta.recordedStepEndEffectChanges.itemTypesBool[type] = true;
+    meta.recordedStepEndEffectChanges.somethingChanged = true;
+    meta.recordedEffectChanges.itemTypesBool[type] = true;
+    meta.recordedEffectChanges.somethingChanged = true;
+  });
+}
+
+export function _addItem({ type, id, state, refs }: { type: string; id: string; state?: any; refs?: any }) {
+  if (!meta.willAddItemsInfo[type]) meta.willAddItemsInfo[type] = {};
+  meta.willAddItemsInfo[type][id] = true;
+  runWhenAddingAndRemovingItems(() => {
+    meta.nowState[type][id] = {
+      ...meta.defaultStateByItemType[type](id),
+      ...(state || {}),
+    };
+    meta.nowRefs[type][id] = {
+      ...meta.defaultRefsByItemType[type](id, meta.nowState[id]),
+      ...(refs || {}),
+    };
+    meta.itemIdsByItemType[type].push(id);
+    meta.recordedStepEndEffectChanges.itemTypesBool[type] = true;
+
+    // TODO Figure out if adding an item should record the properties as changed or not?
+
+    meta.recordedStepEndEffectChanges.itemPropsBool[type][id] = {};
+    meta.recordedEffectChanges.itemPropsBool[type][id] = {};
+
+    meta.diffInfo.propsChanged[type][id] = [];
+    meta.diffInfo.propsChangedBool[type][id] = {};
+
+    meta.recordedStepEndEffectChanges.itemIdsBool[type][id] = true;
+    meta.recordedStepEndEffectChanges.somethingChanged = true;
+    meta.recordedEffectChanges.itemTypesBool[type] = true;
+
+    meta.recordedEffectChanges.itemIdsBool[type][id] = true;
+    meta.recordedEffectChanges.somethingChanged = true;
+
+    // NOTE new items with props different to the defaults props are recorded as changed
+    const itemPropNames = meta.propNamesByItemType[type];
+    forEach(itemPropNames, (propName) => {
+      const propChangedFromDefault =
+        meta.nowState[type][id][propName] !== meta.defaultStateByItemType[type](id)[propName];
+      if (propChangedFromDefault) {
+        meta.recordedStepEndEffectChanges.itemPropsBool[type][id][propName] = true;
+        meta.recordedEffectChanges.itemPropsBool[type][id][propName] = true;
+      }
+    });
+  });
+}
 
 export const getDefaultStates = (): DefaultStates => meta.defaultStateByItemType as DefaultStates;
 
@@ -21,8 +136,9 @@ export const getItemTypes = (): ItemType[] => meta.itemTypeNames;
 
 export const getItemIds = (kind: ItemType): string[] => meta.itemIdsByItemType[kind];
 
-export const getState_OLD = (): DeepReadonly<AllState> => meta.nowState as DeepReadonly<AllState>;
+const _getNestedState = (): DeepReadonly<AllState> => meta.nowState as DeepReadonly<AllState>;
 // export const getState = (kind: string, itemId: string): DeepReadonly<AllState> => meta.nowState as DeepReadonly<AllState>;
+
 export const getState = <T_Kind extends ItemType>(
   kind: T_Kind,
   itemId?: string
@@ -48,15 +164,11 @@ export const getState = <T_Kind extends ItemType>(
 };
 
 // export const setState: SetRepondState<AllState> = (newState) => _setState(newState);
-export const setState_OLD = _setState_OLD;
-export const setState = _setState;
 
 // Good for running things to be sure the state change is seen
 export function onNextTick(callback?: RepondCallback) {
   if (callback) meta.nextTickQueue.push(callback);
 }
-
-export const getPrevState_OLD = (): AllState => meta.prevState as AllState;
 
 export const getPrevState = <T_Kind extends ItemType>(
   kind: T_Kind,
@@ -76,8 +188,6 @@ export const getPrevState = <T_Kind extends ItemType>(
   }
   return meta.prevState[kind][itemId];
 };
-
-export const getRefs_OLD = (): AllRefs => meta.nowRefs as AllRefs;
 
 export const getRefs = <T_Kind extends ItemType>(
   kind: T_Kind,
@@ -132,7 +242,6 @@ export function getItemWillExist<K_Type extends ItemType>(type: K_Type, id: stri
 // Function to selectively get data with only specific props from the repond store, can be used for save data
 export function getPartialState(propsToGet: Partial<ItemPropsByType>) {
   const itemTypes = Object.keys(propsToGet) as Array<keyof ItemPropsByType>;
-  // const state = getState_OLD();
 
   if (!meta.didInit) {
     console.warn("getPartialState called before repond was initialized");
@@ -158,5 +267,5 @@ export function getPartialState(propsToGet: Partial<ItemPropsByType>) {
 }
 
 export function applyState(partialState: Partial<AllState>) {
-  if (partialState) applyPatch(getPatch(getState_OLD(), partialState));
+  if (partialState) applyPatch(getPatch(_getNestedState(), partialState));
 }
